Adding features to golem-common, WIP
