+++ golem-common/Cargo.toml
@@ -12,10 +12,10 @@ description = "Shared code between Golem services"
 harness = false
 
 [dependencies]
-golem-api-grpc = { path = "../golem-api-grpc", version = "0.0.0" }
-golem-wasm-ast = { workspace = true }
-golem-wasm-rpc = { workspace = true }
-golem-rib = { path = "../golem-rib", version = "0.0.0" }
+golem-api-grpc = { path = "../golem-api-grpc", version = "0.0.0", optional = true }
+golem-wasm-ast = { workspace = true, default-features = false }
+golem-wasm-rpc = { workspace = true, default-features = false }
+golem-rib = { path = "../golem-rib", version = "0.0.0", default-features = false }
 
 anyhow = { workspace = true }
 async-trait = { workspace = true }
@@ -38,8 +38,7 @@ lazy_static = { workspace = true }
 poem = { workspace = true }
 poem-openapi = { workspace = true }
 prometheus = { workspace = true }
-prost = { workspace = true }
-prost-types = { workspace = true }
+prost-types = { workspace = true, optional = true }
 rand = { workspace = true }
 range-set-blaze = "0.1.16"
 regex = { workspace = true }
@@ -49,8 +48,8 @@ serde_yaml = { workspace = true }
 sqlx = { workspace = true }
 thiserror = { workspace = true }
 toml = { workspace = true }
+tonic = { workspace = true, optional = true }
 tokio = { workspace = true }
-tonic = { workspace = true }
 tracing = { workspace = true }
 tracing-subscriber = { workspace = true }
 tracing-serde = { workspace = true }
@@ -63,3 +62,8 @@ wasm-wave = { workspace = true }
 anyhow = { workspace = true }
 test-r = { workspace = true }
 tracing-test = { workspace = true }
+
+[features]
+#default = ["grpc"]
+default = []
+grpc = ["dep:golem-api-grpc", "dep:tonic", "dep:prost-types"]
