+++ golem-common/src/retriable_error.rs
@@ -12,8 +12,6 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-use tonic::{Code, Status};
-
 pub trait IsRetriableError {
     /// Returns true if the error is retriable.
     fn is_retriable(&self) -> bool;
@@ -24,32 +22,4 @@ pub trait IsRetriableError {
     /// This is useful to accept some errors as the expected response (such as a not-found result
     /// when looking for a resource).
     fn as_loggable(&self) -> Option<String>;
-}
-
-impl IsRetriableError for Status {
-    fn is_retriable(&self) -> bool {
-        match self.code() {
-            Code::Ok
-            | Code::Cancelled
-            | Code::InvalidArgument
-            | Code::NotFound
-            | Code::AlreadyExists
-            | Code::PermissionDenied
-            | Code::FailedPrecondition
-            | Code::OutOfRange
-            | Code::Unimplemented
-            | Code::DataLoss
-            | Code::Unauthenticated => false,
-            Code::Unknown
-            | Code::DeadlineExceeded
-            | Code::ResourceExhausted
-            | Code::Aborted
-            | Code::Internal
-            | Code::Unavailable => true,
-        }
-    }
-
-    fn as_loggable(&self) -> Option<String> {
-        Some(self.to_string())
-    }
 }
