+++ golem-component-compilation-service/Cargo.toml
@@ -2,6 +2,7 @@
 name = "golem-component-compilation-service"
 version = "0.0.0"
 edition = "2021"
+license = "Apache-2.0"
 
 [lib]
 path = "src/lib.rs"
