+++ golem-component-service-base/Cargo.toml
@@ -2,6 +2,7 @@
 name = "golem-component-service-base"
 version = "0.0.0"
 edition = "2021"
+license = "Apache-2.0"
 
 [lib]
 harness = false
