+++ golem-rib/Cargo.toml
@@ -8,18 +8,16 @@ repository = "https://github.com/golemcloud/golem"
 description = "Parser for Golem's Rib language"
 
 [dependencies]
-golem-api-grpc = { path = "../golem-api-grpc", version = "0.0.0" }
+golem-api-grpc = { path = "../golem-api-grpc", version = "0.0.0", optional = true }
 
-async-trait = { workspace = true }
 bincode = { workspace = true }
 combine = { workspace = true }
-golem-wasm-ast = { workspace = true }
-golem-wasm-rpc = { workspace = true }
-poem-openapi = { workspace = true }
+golem-wasm-ast = { workspace = true, default-features = false }
+golem-wasm-rpc = { workspace = true, default-features = false }
+poem-openapi = { workspace = true, optional = true }
 semver = "1.0.23"
 serde = { workspace = true }
 serde_json = { workspace = true }
-tokio = {workspace = true}
 wasm-wave = { workspace = true }
 
 [dev-dependencies]
@@ -29,3 +27,8 @@ test-r = { workspace = true }
 name = "rib"
 path = "src/lib.rs"
 harness = false
+
+[features]
+default = []
+full = ["dep:golem-api-grpc"]
+poem = ["dep:poem-openapi"]
