+++ golem-rib/src/interpreter/interpreter_stack_value.rs
@@ -17,7 +17,6 @@ use crate::CoercedNumericValue;
 use golem_wasm_ast::analysis::AnalysedType;
 use golem_wasm_rpc::protobuf::type_annotated_value::TypeAnnotatedValue;
 use golem_wasm_rpc::protobuf::typed_result::ResultValue;
-use poem_openapi::types::ToJSON;
 use std::fmt;
 
 // A result of a function can be unit, which is not representable using type_annotated_value
@@ -57,8 +56,8 @@ impl RibInterpreterStackValue {
                 } else {
                     Err(format!(
                         "Unable to complete the math operation on {}, {}",
-                        left.to_json_string(),
-                        right.to_json_string()
+                        serde_json::to_string(&left).map_err(|err| err.to_string())?,
+                        serde_json::to_string(&right).map_err(|err| err.to_string())?
                     ))
                 }
             }
