+++ golem-rib/src/type_inference/rib_input_type.rs
@@ -16,13 +16,13 @@ use crate::{Expr, InferredExpr};
 use bincode::{Decode, Encode};
 use golem_api_grpc::proto::golem::rib::RibInputType as ProtoRibInputType;
 use golem_wasm_ast::analysis::AnalysedType;
-use poem_openapi::Object;
 use serde::{Deserialize, Serialize};
 use std::collections::{HashMap, VecDeque};
 
 // RibInputTypeInfo refers to the required global inputs to a RibScript
 // with its type information. Example: `request` variable which should be of the type `Record`.
-#[derive(Debug, Clone, PartialEq, Serialize, Deserialize, Encode, Decode, Object)]
+#[derive(Debug, Clone, PartialEq, Serialize, Deserialize, Encode, Decode)]
+#[cfg_attr(feature = "poem", derive(poem_openapi::Object))]
 pub struct RibInputTypeInfo {
     pub types: HashMap<String, AnalysedType>,
 }
