+++ golem-worker-executor-base/Cargo.toml
@@ -89,7 +89,6 @@ warp = { workspace = true }
 wasmtime = { workspace = true }
 wasmtime-wasi = { workspace = true }
 wasmtime-wasi-http = { workspace = true }
-windows-sys = "0.52.0"
 zstd = "0.13"
 sqlx = { workspace = true }
