+++ golem-worker-service/Cargo.toml
@@ -52,8 +52,6 @@ regex = { workspace = true }
 serde = { workspace = true }
 serde_json = { workspace = true }
 serde_yaml = { workspace = true }
-strum = { workspace = true }
-strum_macros = { workspace = true }
 tap = { workspace = true }
 tokio = { workspace = true }
 tokio-stream = { workspace = true }
